+++ Prosjekt/Screen.h
@@ -1,16 +1,38 @@
 #pragma once
 
-
 class SpaceDefender;  // Forward declaration of SpaceDefender to avoid circular dependency
 
 // Abstract base class for different screens
 class Screen {
 public:
     virtual ~Screen() = default;  // Virtual destructor to ensure proper cleanup
-
-    // Virtual method to draw the screen's content
     virtual void draw(SpaceDefender& window) = 0;
+};
 
-    // Virtual method to handle user input (e.g., button clicks, key presses)
-    virtual void handleInput(SpaceDefender& window) = 0;
+
+// ---- ScreenMenu ----
+class ScreenMenu : public Screen {
+public:
+    void draw(SpaceDefender& window) override;
+};
+
+
+// ---- ScreenGame ----
+class ScreenGame : public Screen {
+public:
+    void draw(SpaceDefender& window) override;
+};
+
+
+// ---- ScreenHighscore ----
+class ScreenHighscore : public Screen {
+public:
+    void draw(SpaceDefender& window) override;
+};
+        
+
+// ---- ScreenSettings ----
+class ScreenSettings : public Screen {
+public:
+    void draw(SpaceDefender& window) override;
 };
