--- Prosjekt/ScreenHighscore.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "SpaceDefender.h"
-#include "Screen.h"  // Include the base Screen class
-
-class ScreenHighscore : public Screen {
-public:
-    void draw(SpaceDefender& window) override;
-    void handleInput(SpaceDefender& window) override;
-};
-
