--- Prosjekt/ScreenSettings.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "SpaceDefender.h"
-#include "Screen.h"  // Include the base Screen class
-
-class ScreenSettings : public Screen {
-public:
-    void draw(SpaceDefender& window) override;
-    void handleInput(SpaceDefender& window) override;
-};
